requirements.txt
PyQt5==5.15.6
Pillow==8.4.0
PyPDF2==1.26.0
psutil==5.8.0
